Fixing bug

+++ src/main/java/com/ikaragil/baksos/service/impl/BaksosServiceImpl.java
@@ -1,6 +1,7 @@
 package com.ikaragil.baksos.service.impl;
 
 import com.ikaragil.baksos.domain.Baksos;
+import com.ikaragil.baksos.domain.BaksosDetail;
 import com.ikaragil.baksos.domain.Search;
 import com.ikaragil.baksos.repository.impl.BaksosDaoImpl;
 import com.ikaragil.baksos.service.BaksosService;
@@ -80,6 +81,26 @@ public class BaksosServiceImpl implements BaksosService {
     }
 
     @Override
+    public Map getDetail(Search search) {
+        Map m = new HashMap();
+        if (search.getKey().equals("kode")) {
+            List<BaksosDetail> details = baksosDao.getDetail(search.getValue());
+            if (details.isEmpty()) {
+                m.put("kode", "01");
+                m.put("response", "Kode Tidak Ditemukan");
+            } else {
+                m.put("data", details);
+                m.put("kode", "00");
+                m.put("response", "Data Ditemukan");
+            }
+        } else {
+            m.put("kode", "02");
+            m.put("response", "Keyword Salah");
+        }
+        return m;
+    }
+
+    @Override
     public Map findAll() {
         Map m = new HashMap();
         List<Baksos> findAll = baksosDao.findAll();
@@ -98,6 +119,14 @@ public class BaksosServiceImpl implements BaksosService {
     public Map create(Baksos baksos) {
         Map m = new HashMap();
         if (baksos.getId() == null) {
+            List<String> lastId = baksosDao.getLastId();
+            String tgl = baksos.getTgl_baksos().substring(0, 6);
+            if (lastId.size() == 0) {
+                String id = "0";
+                baksos.setKode("IKA" + tgl + "-" + (Integer.parseInt(id) + 1));
+            } else {
+                baksos.setKode("IKA" + tgl + "-" + (Integer.parseInt(lastId.get(0)) + 1));
+            }
             m.put("data", baksosDao.create(baksos));
             m.put("kode", "00");
             m.put("response", "Sukses");
@@ -113,9 +142,15 @@ public class BaksosServiceImpl implements BaksosService {
     public Map update(Baksos baksos) {
         Map m = new HashMap();
         if (baksos.getId() != null) {
-            m.put("data", baksosDao.update(baksos));
-            m.put("kode", "00");
-            m.put("response", "Sukses");
+            List<Baksos> findById = baksosDao.findById(baksos.getId());
+            if (findById.isEmpty()) {
+                m.put("kode", "01");
+                m.put("response", "Id not found");
+            } else {
+                m.put("data", baksosDao.update(baksos));
+                m.put("kode", "00");
+                m.put("response", "Sukses");
+            }
         } else {
             m.put("kode", "01");
             m.put("response", "Gagal");
@@ -127,9 +162,15 @@ public class BaksosServiceImpl implements BaksosService {
     public Map delete(Baksos baksos) {
         Map m = new HashMap();
         if (baksos.getId() != null) {
-            m.put("data", baksosDao.delete(baksos));
-            m.put("kode", "00");
-            m.put("response", "Sukses");
+            List<Baksos> findById = baksosDao.findById(baksos.getId());
+            if (findById.isEmpty()) {
+                m.put("kode", "01");
+                m.put("response", "Id not found");
+            } else {
+                m.put("data", baksosDao.delete(baksos));
+                m.put("kode", "00");
+                m.put("response", "Sukses");
+            }
         } else {
             m.put("kode", "01");
             m.put("response", "Gagal");
@@ -141,7 +182,7 @@ public class BaksosServiceImpl implements BaksosService {
     public Map findById(Search search) {
         Map m = new HashMap();
         if (search.getKey().equals("id")) {
-            List<Baksos> findById = baksosDao.findById(Integer.parseInt(search.getValue()));
+            List<Baksos> findById = baksosDao.findById(Long.parseLong(search.getValue()));
             if (findById.isEmpty()) {
                 m.put("kode", "01");
                 m.put("response", "Data Tidak Ditemukan");
